sql_agent_service/requirements.txt
fastapi
uvicorn[standard]
pymysql
sqlalchemy
google-generativeai>=0.5.0
google-cloud-aiplatform>=1.49.0
cloud-sql-python-connector[pymysql]
coloredlogs
python-dotenv
sse-starlette